에너미 세력 3종 데이터 생성

+++ Assets/Scripts/Battle/AttackableUnit.cs
@@ -62,9 +62,9 @@ public abstract class AttackableUnit : MonoBehaviour
     [SerializeField, Header("에너미 리스트")] protected List<AttackableUnit> enemyList;
     [SerializeField, Header("시민 리스트")] protected List<AttackableUnit> citizenList;
 
-    public int MaxHp => (int)((bufferState.maxHealthIncrease * characterData.data.healthPoint) + characterData.data.healthPoint);
-    public float UnitHpScale => (float)characterData.data.currentHp / MaxHp;
-    public virtual int UnitHp {
+    public float MaxHp => ((bufferState.maxHealthIncrease * characterData.data.healthPoint) + characterData.data.healthPoint);
+    public float UnitHpScale => characterData.data.currentHp / MaxHp;
+    public virtual float UnitHp {
         get { return characterData.data.currentHp; }
         set {
             characterData.data.currentHp = Mathf.Clamp(value, 0, MaxHp);
