+++ Assets/Scripts/UI/BattleUi/HeroUi.cs
@@ -41,9 +41,9 @@ public class HeroUi : MonoBehaviour
     {
         viewBuffList.OnClickPopUp();
     }
-    public void SetHp(int nowHp, int maxHp)
+    public void SetHp(float nowHp, float maxHp)
     {
-        hpBar.value = (float)nowHp / maxHp; // Mathf.Max((float)nowHp / maxHp); max 왜 쓴거임? - 진석 
+        hpBar.value = nowHp / maxHp; // Mathf.Max((float)nowHp / maxHp); max 왜 쓴거임? - 진석 
                                             // 그러게요..? 여기다 쓸 이유가 없는데.. 최대체력 넘지 못하게 하는 코드
                                             //  작업하면서 무지성으로 복붙한거 같아요- 정연
     }
